latex2mobi-converter/src/main/java/at/ac/tuwien/ims/latex2mobiformulaconv/converter/mathml2html/elements/tablesmatrices/Mtable.java
package at.ac.tuwien.ims.latex2mobiformulaconv.converter.mathml2html.elements.tablesmatrices;

import at.ac.tuwien.ims.latex2mobiformulaconv.converter.mathml2html.elements.FormulaElement;
import at.ac.tuwien.ims.latex2mobiformulaconv.converter.mathml2html.elements.layout.Mfenced;
import org.jdom2.Element;

import java.util.ArrayList;
import java.util.List;

/**
 * The MIT License (MIT)
 * latex2mobi -- LaTeX Formulas to Mobi Converter
 * Copyright (c) 2014 Michael Auß
 * <p/>
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 * <p/>
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 * <p/>
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 * <p/>
 * <p/>
 * For Third Party Software Licenses read LICENSE file in base dir.
 *
 * Renders a MathML Mtable tag (Matrices, tables)
 *
 * http://www.w3.org/TR/MathML2/chapter3.html#presm.mtable
 *
 * @author Michael Auß
 *         Created: 20.05.14 23:36
 */
public class Mtable implements FormulaElement {
    private List<Mtr> rows = new ArrayList<Mtr>();

    public List<Mtr> getRows() {
        return rows;
    }

    @Override
    public Element render(FormulaElement parent, List<FormulaElement> siblings) {
        Element mtableDiv = new Element("div");
        mtableDiv.setAttribute("class", "mtable");

        // create Table
        Element table = new Element("table");

        // Matrix / Table parenthesis
        if (parent != null && parent instanceof Mfenced) {
            Mfenced mfenced = (Mfenced) parent;
            if (mfenced.getOpened().equals("(") && mfenced.getClosed().equals(")")) {
                table.setAttribute("class", "pmatrix");
            }
            if (mfenced.getOpened().equals("[") && mfenced.getClosed().equals("]")) {
                table.setAttribute("class", "bmatrix");
            }
            if (mfenced.getOpened().equals("{") && mfenced.getClosed().equals("}")) {
                table.setAttribute("class", "pmatrix"); // intentionally pmatrix for curved border corners
                mtableDiv.setAttribute("class", mtableDiv.getAttributeValue("class") + " mtable-Bmatrix");
            }
            if (mfenced.getOpened().equals("|") && mfenced.getClosed().equals("|")) {
                table.setAttribute("class", "vmatrix");
            }
            if (mfenced.getOpened().equals("∥") && mfenced.getClosed().equals("∥")) {
                table.setAttribute("class", "Vmatrix");
            }

        }

        // evaluate Rows
        for (int i = 0; i < rows.size(); i++) {
            table.addContent(rows.get(i).render(null, null));
        }

        mtableDiv.addContent(table);

        return mtableDiv;
    }
}
